feat(be-65):(настройка проекта) структурирование тестов
