+++ src/interviews/__tests__/__unit__/utils/generatePrefixOfInterview.spec.ts
